+++ tests/Amadeus/ClientTest.php
@@ -527,6 +527,62 @@ class ClientTest extends BaseTestCase
         $this->assertEquals($messageResult, $response);
     }
 
+    public function testCanDoDummyPnrTransferOwnershipCall()
+    {
+        $mockSessionHandler = $this->getMockBuilder('Amadeus\Client\Session\Handler\HandlerInterface')->getMock();
+
+        $mockedSendResult = new Client\Session\Handler\SendResult();
+        $mockedSendResult->responseObject = new \stdClass();
+        $mockedSendResult->responseObject->dummyProp = 'A dummy message result'; // Not an actual Soap reply.
+        $mockedSendResult->responseXml = 'A dummy message result'; // Not an actual XML reply
+
+        $messageResult = new Client\Result($mockedSendResult);
+
+        $expectedPnrResult = new Client\Struct\Pnr\TransferOwnership(
+            new Client\RequestOptions\PnrTransferOwnershipOptions([
+                'recordLocator' => 'ABC123',
+                'newOffice' => 'BRUXXXXXX'
+            ])
+        );
+
+        $mockSessionHandler
+            ->expects($this->once())
+            ->method('sendMessage')
+            ->with('PNR_TransferOwnership', $expectedPnrResult, ['endSession' => false])
+            ->will($this->returnValue($mockedSendResult));
+        $mockSessionHandler
+            ->expects($this->once())
+            ->method('getMessagesAndVersions')
+            ->will($this->returnValue(['PNR_TransferOwnership' => '14.1']));
+
+        $mockResponseHandler = $this->getMockBuilder('Amadeus\Client\ResponseHandler\ResponseHandlerInterface')->getMock();
+
+        $mockResponseHandler
+            ->expects($this->once())
+            ->method('analyzeResponse')
+            ->with($mockedSendResult, 'PNR_TransferOwnership')
+            ->will($this->returnValue($messageResult));
+
+        $par = new Params();
+        $par->sessionHandler = $mockSessionHandler;
+        $par->requestCreatorParams = new Params\RequestCreatorParams([
+            'receivedFrom' => 'some RF string',
+            'originatorOfficeId' => 'BRUXXXXXX'
+        ]);
+        $par->responseHandler = $mockResponseHandler;
+
+        $client = new Client($par);
+
+        $response = $client->pnrTransferOwnership(
+            new Client\RequestOptions\PnrTransferOwnershipOptions([
+                'recordLocator' => 'ABC123',
+                'newOffice' => 'BRUXXXXXX'
+            ])
+        );
+
+        $this->assertEquals($messageResult, $response);
+    }
+
 
     public function testCanDoDummyQueueListCall()
     {
@@ -2271,7 +2327,7 @@ class ClientTest extends BaseTestCase
 
         $client = new Client($par);
 
-        $client->airFlightInfo();
+        $client->airFlightInfo(new Client\RequestOptions\AirFlightInfoOptions());
     }
 
     public function testCanSendDocIssuanceIssueTicket()